Moving everything into the original filenames.
Need to stop being so overly cautious.
